OR-1081 clean code
